test(green): spike3 basic command works
Flags and dashdash need a lot more testing. Dashdash doesn't seem
to be working correctly yet.
