+++ deno.lock
@@ -16,6 +16,7 @@
       "jsr:@std/internal@^1.0.3": "jsr:@std/internal@1.0.3",
       "jsr:@std/io@^0.224.9": "jsr:@std/io@0.224.9",
       "jsr:@std/path@^1.0.4": "jsr:@std/path@1.0.6",
+      "jsr:@std/streams@^1.0.6": "jsr:@std/streams@1.0.6",
       "jsr:@std/testing@^1.0.2": "jsr:@std/testing@1.0.2",
       "npm:@sinclair/typebox@0.32.31": "npm:@sinclair/typebox@0.32.31",
       "npm:@sinclair/typebox@^0.33.11": "npm:@sinclair/typebox@0.33.11"
@@ -75,6 +76,12 @@
       },
       "@std/path@1.0.6": {
         "integrity": "ab2c55f902b380cf28e0eec501b4906e4c1960d13f00e11cfbcd21de15f18fed"
+      },
+      "@std/streams@1.0.6": {
+        "integrity": "022ed94e380d06b4d91c49eb70241b7289ab78b8c2b4c4bbb7eb265e4997c25c",
+        "dependencies": [
+          "jsr:@std/bytes@^1.0.2"
+        ]
       },
       "@std/testing@1.0.2": {
         "integrity": "9e8a7f4e26c219addabe7942d09c3450fa0a74e9662341961bc0ef502274dec3",
@@ -154,6 +161,7 @@
       "jsr:@std/cli@^1.0.6",
       "jsr:@std/expect@^1.0.3",
       "jsr:@std/io@^0.224.9",
+      "jsr:@std/streams@^1.0.6",
       "jsr:@std/testing@^1.0.2",
       "npm:@sinclair/typebox@^0.33.11"
     ]
